1) Updates for GR 3.7.0git.  This should compile against both recent 3.6 (tested against 3.6.4.2) and current 3.7.
This makes things a bit ugly, as the main cmake file now both looks for
gnuradio-core (which is an indicator of 3.6) and tries the new 3.7 way
of locating GR.  The results are chosen and passed on to other cmake
files, and a new macro (GR_OLD) should be defined iff 3.6 is detected.

2) In pocencode_impl.cc, return the number of bytes actually
transferred fro work().